requirements.txt
torch>=2.1
torchvision>=0.16
timm>=1.0.3
scikit-learn>=1.4
pandas>=2.2
numpy>=1.26
Pillow>=10.0
